Add solutions for day 10 and 11, change result type from int to long

+++ AoC2022/Inputs/AbstractInput.cs
@@ -2,17 +2,17 @@ namespace Inputs;
 
 public abstract class AbstractInput<TU, TV> : IInput
 {
-    public int FirstPuzzle(string input) => SolveFirstPuzzle(ParseInput(input));
+    public long FirstPuzzle(string input) => SolveFirstPuzzle(ParseInput(input));
 
-    public int SecondPuzzle(string input) => SolveSecondPuzzle(ParseInputTwo(input));
+    public long SecondPuzzle(string input) => SolveSecondPuzzle(ParseInputTwo(input));
 
     protected abstract TU ParseInput(string input);
 
     protected abstract TV ParseInputTwo(string input);
 
-    protected abstract int SolveFirstPuzzle(TU input);
+    protected abstract long SolveFirstPuzzle(TU input);
 
-    protected abstract int SolveSecondPuzzle(TV input);
+    protected abstract long SolveSecondPuzzle(TV input);
 
     public abstract string TestInput { get; }
     public abstract string RealInput { get; }
